#163 - Extract classes in package 'de.tudarmstadt.ukp.wikipedia.parser.tutorial' into a separate Maven module
- Separates demonstration classes that existed in 'de.tudarmstadt.ukp.wikipedia.parser'
- Introduces new maven module 'de.tudarmstadt.ukp.wikipedia.tutorial'
- Moves classes T1_... to T6_... and HtmlFileDemo to the new module
- Improves JavaDoc in HtmlWriter
