+++ de.tudarmstadt.ukp.wikipedia.tutorial/src/main/java/de/tudarmstadt/ukp/wikipedia/tutorial/T1_SimpleParserDemo.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package de.tudarmstadt.ukp.wikipedia.parser.tutorial;
+package de.tudarmstadt.ukp.wikipedia.tutorial;
 
 import java.io.IOException;
